requirements.txt
Django>=5.0.0
hdfs>=2.6.0
pyhive>=0.6.1
requests>=2.25.1
djangorestframework>=3.12.0
psycopg2-binary>=2.9.1
pandas>=1.3.0
numpy>=1.21.0
celery>=5.2.0
django-celery-beat>=2.2.0
django-celery-results>=2.2.0
redis>=3.5.3
